+++ app/controllers/api/v1/merchants_controller.rb
@@ -6,9 +6,18 @@ class Api::V1::MerchantsController < Api::V1::BaseController
   def create
     merchant = Merchant.new(merchant_params)
     if merchant.save
-      render json: merchant.as_json
+      render json: { merchant_id: merchant.id }
     else
       respond_with_errors "While creating new Merchant: ", 422, merchant.errors.messages
+    end
+  end
+
+  def verify_otp
+    @merchant = Merchant.find(params[:id])
+    if @merchant.valid_otp?(params[:otp])
+      render json: @merchant.as_json
+    else
+      head :bad_request
     end
   end
