+++ app/views/api/v1/admin/merchants/_merchant.json.jbuilder
@@ -1,1 +1,3 @@
+# frozen_string_literal: true
+
 json.merge! merchant.as_json
