+++ app/views/api/v1/admin/merchants/index.json.jbuilder
@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 json.merchants @merchants do |merchant|
-  json.partial! 'merchant', merchant: merchant
+  json.partial! "merchant", merchant: merchant
 end
