+++ db/migrate/20200728083700_add_temp_id_for_merchants.rb
@@ -1,0 +1,5 @@
+class AddTempIdForMerchants < ActiveRecord::Migration[6.0]
+  def change
+    add_column :merchants, :temp_id, :text 
+  end
+end
