+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_07_28_080517) do
+ActiveRecord::Schema.define(version: 2020_07_28_083700) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "pgcrypto"
@@ -70,6 +70,7 @@ ActiveRecord::Schema.define(version: 2020_07_28_080517) do
     t.datetime "updated_at", precision: 6, null: false
     t.string "lb_code"
     t.text "otp_token"
+    t.text "temp_id"
   end
 
   create_table "qr_codes", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
